extract mobile header
